fix(tools): split estimateDurationTool out to fix hard-cap violation
convex/ai/tools.ts reached 401 lines after merging main (which added
the estimate_duration tool). Move estimateDurationTool to its own
convex/ai/estimationTools.ts so tools.ts stays under the 400-line hard
cap enforced by CI.

No behaviour change — coach.ts now imports estimateDurationTool from
the new file instead of tools.ts.
